stage-1/requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cu121

# numpy
# torch==2.2.2+cu121
# torchvision==0.17.2+cu121
# torchaudio==2.2.2+cu121

# yt-dlp
# # whisper @ git+https://github.com/openai/whisper.git #@c0d2f6241a3819c13a3fa58f41edfdc0c5b44b7e
# faster-whisper==1.1.1
# # ollama==0.5.1
# llama_cpp_python==0.3.9
# huggingface_hub==0.33.1



# Repositório extra para PyTorch com suporte a CUDA 12.1
--extra-index-url https://download.pytorch.org/whl/cu121

# Repositório extra para Llama CPP com suporte a CUDA 12.1
--extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu121

# Pacotes PyTorch
torch
torchvision
torchaudio

# Pacotes da Aplicação
yt-dlp
faster-whisper
llama-cpp-python
huggingface-hub